Abort harder when config processing fails.
It turns out that silently ignoring unexpected config failures is bad! Who
knew. Fortunately, pkg_resources raises an exception when this happens. By
special-casing the one expected scenario (failure to load a connector config
when scannning configurers, since connectors are not configurers) the
built-in exception handling can report the failure.